move tracker to a common package
Gateway API will use a changing tracker in a future version, the preview
will always parse all manifests whenever a Gateway related change is
made. Maybe the tracking strategy changes, but currently some core
functionalities need a tracking object instantiated, eg get secret. So
moving to minimally use the current implementation in order to impact
shared code as few as possible.
